% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Implementation of the Dijkstra algorithm
The algorithm takes a graph and an initial node 
and calculates the shortest path from the initial 
node to every other node in the graph.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Graph with weights}

\item{init_node}{Starting node}
}
\value{
shortest path to every other node
}
\description{
Pseudocode from: https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\examples{
dijkstra(graph, 1)
}
