% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find the greatest common divisor of two numbers.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The greatest common divisor
}
\description{
Pseudocode from: https://en.wikipedia.org/wiki/Euclidean_algorithm.
}
\examples{
euclidean(100, 1000)
}
